requirements.txt
streamlit==1.42.0
pandas==2.2.0
numpy==1.26.4
google-cloud-bigquery==3.20.0
scipy==1.13.0
py_vollib==1.0.1
requests==2.31.0